+++ pulsar-common/src/main/java/org/apache/pulsar/common/tls/PublicSuffixList.java
@@ -21,7 +21,7 @@
  * From Apache HTTP client
  */
 
-package org.apache.pulsar.client.impl.tls;
+package org.apache.pulsar.common.tls;
 
 import java.util.Collections;
 import java.util.List;
